Merge branch 0.9 into master

+++ fev/fev_conn.c
@@ -87,8 +87,8 @@ int    fev_conn(fev_state* fev,
         fev_conn_info* conn_info = calloc(1, sizeof(fev_conn_info));
         conn_info->fd = sockfd;
         fev_timer_svc* timer_svc = fev_module_data(fev, FEV_CONN_MODULE_NAME);
-        conn_info->timer = fev_tmsvc_timer_add(timer_svc, (uint32_t)timeout,
-                                                on_timer, conn_info);
+        conn_info->timer = fev_tmsvc_timer_add(timer_svc, timeout,
+                                               on_timer, conn_info);
         if (!conn_info->timer) {
             close(sockfd);
             free(conn_info);
